requirements.txt
duckdb==0.10.3
chromadb==1.5.8
groq==0.9.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
sentence-transformers==4.1.0
torch==2.7.0
pandas==2.2.3
networkx==3.3
scikit-learn==1.6.1
numpy==2.2.6
